% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PhysioNeurophys-package}
\alias{PhysioNeurophys}
\alias{PhysioNeurophys-package}
\title{PhysioNeurophys: TMS and peripheral motor-neurophysiology analysis}
\description{
Tools for TMS pulse metadata, MEP amplitude and latency, recruitment curves,
H-reflex and F-wave analysis, paired-pulse SICI/ICF, cortical silent-period
timing, source-informed TMS-EEG cleaning, TMS-evoked-potential summaries,
and PREP2 upper-limb prognosis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioNeurophys}
  \item \url{https://x-biosignal.r-universe.dev/PhysioNeurophys}
  \item \url{https://x-biosignal.github.io/PhysioNeurophys/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioNeurophys/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
